I removed the doc folder with .texi files and replaced it with the pdf of the documentation. I could not find a reliable way to view .texi files.
I used makeinfo --html (mac) on one of the .texi files and generated
.html files to view, but there also seemed to be formatting errors,
and texinfo seems to have been discontinued according to a stackoverflow
post (sorry, no link).

We may have to use .texi in the future in which case a better way might
be shown to view these files, but in the meantime it just seemed better
to get rid of these files.